requirements.txt
Django>=5.2,<5.3
whitenoise>=6.7
django-cors-headers>=4.4
opencv-python>=4.10
pytesseract>=0.3.10
Pillow>=10.0
gunicorn>=22.0